Move peerlistv2 into peerlist/v2
