+++ api/peer/list.go
@@ -71,7 +71,7 @@ type ChooserList interface {
 // peerlist.List and ListImplementation compose well with sharding schemes the
 // degenerate to returning the only available peer.
 //
-// Deprecated in favor of "go.uber.org/yarpc/peer/peerlistv2".Implementation.
+// Deprecated in favor of "go.uber.org/yarpc/peer/peerlist/v2".Implementation.
 type ListImplementation interface {
 	transport.Lifecycle
