+++ peer/pendingheap/heap.go
@@ -27,7 +27,7 @@ import (
 
 	"go.uber.org/yarpc/api/peer"
 	"go.uber.org/yarpc/api/transport"
-	peerlist "go.uber.org/yarpc/peer/peerlistv2"
+	"go.uber.org/yarpc/peer/peerlist/v2"
 )
 
 type pendingHeap struct {
